requirements.txt
fastapi==0.110.1
python-dotenv==1.0.1
Requests==2.31.0
uvicorn==0.29.0
